+++ packages/react/src/components/alert/alert.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryFn } from "@storybook/react-vite"
 import { PropsTable } from "../../../storybook/components"
 import { COLOR_SCHEMES, noop } from "../../utils"
-import { CloseButton } from "../button"
+import { CloseButton } from "../close-button"
 import { VStack } from "../stack"
 import { Alert } from "./"
