+++ packages/react/src/components/button/index.ts
@@ -3,10 +3,3 @@ export type { ButtonProps } from "./button"
 export { ButtonGroup } from "./button-group"
 export type { ButtonGroupProps } from "./button-group"
 export * from "./button.style"
-export {
-  CloseButton,
-  CloseButtonPropsContext,
-  useCloseButtonPropsContext,
-} from "./close-button"
-export type { CloseButtonProps } from "./close-button"
-export * from "./close-button.style"
