+++ packages/react/src/components/close-button/close-button.style.ts
