+++ packages/react/src/components/close-button/index.ts
@@ -1,0 +1,12 @@
+export type { ButtonProps } from "../button"
+export { Button, ButtonPropsContext, useButtonPropsContext } from "../button"
+export { ButtonGroup } from "../button"
+export type { ButtonGroupProps } from "../button"
+export * from "../button/button.style"
+export {
+  CloseButton,
+  CloseButtonPropsContext,
+  useCloseButtonPropsContext,
+} from "./close-button"
+export type { CloseButtonProps } from "./close-button"
+export * from "./close-button.style"
