+++ packages/react/src/components/input/input.stories.tsx
@@ -4,10 +4,11 @@ import { useForm } from "react-hook-form"
 import { withMask } from "use-mask-input"
 import { PropsTable } from "../../../storybook/components"
 import { COLOR_SCHEMES, toTitleCase } from "../../utils"
-import { Button, IconButton } from "../button"
+import { Button } from "../button"
 import { Field } from "../field"
 import { For } from "../for"
 import { MailIcon, PhoneIcon, SearchIcon } from "../icon"
+import { IconButton } from "../icon-button"
 import { NativeSelect } from "../native-select"
 import { VStack } from "../stack"
 import { Input, InputGroup } from "./"
