+++ packages/react/src/index.ts
@@ -23,6 +23,7 @@ export * from "./components/card"
 export * from "./components/carousel"
 export * from "./components/center"
 export * from "./components/circle-progress"
+export * from "./components/close-button"
 export * from "./components/code"
 export * from "./components/collapse"
 export * from "./components/color-swatch"
